infrastructure/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
python-dotenv>=0.19.0
boto3>=1.26.0